Taller22/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-storage-blob==12.19.0
azure-search-documents==11.4.0b9
azure-identity
openai[datalib]==0.27.6
tiktoken
langchain
networkx==3.2.1
unstructured[pdf]==0.11.0
protobuf==3.19.4